+++ src/utilities.js
@@ -194,10 +194,20 @@ export const getFirebaseDoc = (firebaseRef, schema, versionNumber) => {
 				.startAt(String(mostRecentRemoteKey + 1))
 				.endAt(String(versionNumber))
 				.once('value');
+			const getLatestChange = firebaseRef
+				.child('changes')
+				.orderByKey()
+				.limitToLast(1)
+				.once('value');
+			const getLatestMerge = firebaseRef
+				.child('merges')
+				.orderByKey()
+				.limitToLast(1)
+				.once('value');
 
-			return Promise.all([newDoc, getChanges, getMerges]);
+			return Promise.all([newDoc, getChanges, getMerges, getLatestChange, getLatestMerge]);
 		})
-		.then(([newDoc, changesSnapshot, mergesSnapshot]) => {
+		.then(([newDoc, changesSnapshot, mergesSnapshot, latestChange, latestMerge]) => {
 			const changesSnapshotVal = changesSnapshot.val() || {};
 			const mergesSnapshotVal = mergesSnapshot.val() || {};
 			const allKeyables = { ...changesSnapshotVal, ...mergesSnapshotVal };
@@ -205,6 +215,13 @@ export const getFirebaseDoc = (firebaseRef, schema, versionNumber) => {
 			const stepClientIds = [];
 			const keys = Object.keys(allKeyables);
 			mostRecentRemoteKey = keys.length ? Math.max(...keys) : mostRecentRemoteKey;
+
+			const latestKey = Object.keys({
+				...latestChange.val(),
+				...latestMerge.val(),
+			})
+				.map((key) => parseInt(key, 10))
+				.reduce((max, next) => Math.max(max, next), 0);
 
 			/* flattenedMergeStepArray is an array of { steps, client, time } values */
 			/* It flattens the case where we have a merge-object which is an array of */
@@ -215,6 +232,11 @@ export const getFirebaseDoc = (firebaseRef, schema, versionNumber) => {
 				}
 				return [...prev, allKeyables[curr]];
 			}, []);
+
+			const latestTimestamp =
+				flattenedMergeStepArray.length > 0
+					? flattenedMergeStepArray[flattenedMergeStepArray.length - 1].t
+					: null;
 
 			/* Uncompress steps and add stepClientIds */
 			flattenedMergeStepArray.forEach((stepContent) => {
@@ -247,6 +269,13 @@ export const getFirebaseDoc = (firebaseRef, schema, versionNumber) => {
 			return {
 				content: updatedDoc.toJSON(),
 				mostRecentRemoteKey: mostRecentRemoteKey,
+				historyData: {
+					timestamps: {
+						[versionNumber]: latestTimestamp,
+					},
+					currentKey: Number(versionNumber) || latestKey,
+					latestKey: latestKey,
+				},
 			};
 		})
 		.catch((firebaseErr) => {