+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- update: collaborative plugin now fires an onUpdateLatestKey callback with the latest remote
+key from Firebase.
 - update: BREAKING - collaborationOptions updated to take ref, rather than firebase credentials
 - update: BREAKING - new firebase schema to support merges, branches, and other metadata.